Added a new auto
