--- src/main/deploy/pathplanner/autos/Amp Blue Side Auto Regular.auto
@@ -1,79 +1,0 @@
-{
-  "version": 1.0,
-  "startingPose": {
-    "position": {
-      "x": 0.7509152340242001,
-      "y": 6.595503105536126
-    },
-    "rotation": 58.46520809481169
-  },
-  "command": {
-    "type": "sequential",
-    "data": {
-      "commands": [
-        {
-          "type": "named",
-          "data": {
-            "name": "TransitShootSpeaker"
-          }
-        },
-        {
-          "type": "named",
-          "data": {
-            "name": "IntakeTransitCommand"
-          }
-        },
-        {
-          "type": "path",
-          "data": {
-            "pathName": "Amp Side Blue trajectory1"
-          }
-        },
-        {
-          "type": "path",
-          "data": {
-            "pathName": "Amp Side Blue trajectory2"
-          }
-        },
-        {
-          "type": "wait",
-          "data": {
-            "waitTime": 0.95
-          }
-        },
-        {
-          "type": "path",
-          "data": {
-            "pathName": "Amp Side Blue Regular trajectory3"
-          }
-        },
-        {
-          "type": "wait",
-          "data": {
-            "waitTime": 0.95
-          }
-        },
-        {
-          "type": "named",
-          "data": {
-            "name": "IntakeTransitStopCommand"
-          }
-        },
-        {
-          "type": "named",
-          "data": {
-            "name": "TransitShootSpeaker"
-          }
-        },
-        {
-          "type": "wait",
-          "data": {
-            "waitTime": 0.95
-          }
-        }
-      ]
-    }
-  },
-  "folder": null,
-  "choreoAuto": false
-}
