--- src/main/deploy/pathplanner/autos/Non-Amp Blue Side Auto.auto
@@ -1,67 +1,0 @@
-{
-  "version": 1.0,
-  "startingPose": {
-    "position": {
-      "x": 0.4234789096358818,
-      "y": 4.186507290393498
-    },
-    "rotation": 2.4366482468101056
-  },
-  "command": {
-    "type": "sequential",
-    "data": {
-      "commands": [
-        {
-          "type": "path",
-          "data": {
-            "pathName": "Non-Amp Blue Side trajectory 0.2"
-          }
-        },
-        {
-          "type": "named",
-          "data": {
-            "name": "TransitShootSpeaker"
-          }
-        },
-        {
-          "type": "named",
-          "data": {
-            "name": "IntakeTransitCommand"
-          }
-        },
-        {
-          "type": "path",
-          "data": {
-            "pathName": "Non-Amp Blue Side trajectory 0.3"
-          }
-        },
-        {
-          "type": "path",
-          "data": {
-            "pathName": "New Non-Amp Blue Side trajectory 0.5"
-          }
-        },
-        {
-          "type": "wait",
-          "data": {
-            "waitTime": 0.95
-          }
-        },
-        {
-          "type": "path",
-          "data": {
-            "pathName": "Non-Amp Blue Side trajectory 1"
-          }
-        },
-        {
-          "type": "path",
-          "data": {
-            "pathName": "Non Amp Blue Side trajectory 2"
-          }
-        }
-      ]
-    }
-  },
-  "folder": null,
-  "choreoAuto": false
-}
