--- src/main/deploy/pathplanner/paths/Amp Side Blue Regular trajectory3.path
@@ -1,70 +1,0 @@
-{
-  "version": 1.0,
-  "waypoints": [
-    {
-      "anchor": {
-        "x": 1.2303755661642377,
-        "y": 6.373314171129768
-      },
-      "prevControl": null,
-      "nextControl": {
-        "x": 6.013284733122172,
-        "y": 6.759221267730286
-      },
-      "isLocked": false,
-      "linkedName": null
-    },
-    {
-      "anchor": {
-        "x": 8.749716872653117,
-        "y": 7.449176379834242
-      },
-      "prevControl": {
-        "x": 4.668456972241579,
-        "y": 7.472564688719121
-      },
-      "nextControl": null,
-      "isLocked": false,
-      "linkedName": null
-    }
-  ],
-  "rotationTargets": [],
-  "constraintZones": [],
-  "eventMarkers": [
-    {
-      "name": "New Event Marker",
-      "waypointRelativePos": 0.35,
-      "command": {
-        "type": "sequential",
-        "data": {
-          "commands": [
-            {
-              "type": "named",
-              "data": {
-                "name": "IntakeTransitCommand"
-              }
-            }
-          ]
-        }
-      }
-    }
-  ],
-  "globalConstraints": {
-    "maxVelocity": 6.0,
-    "maxAcceleration": 3.0,
-    "maxAngularVelocity": 360.0,
-    "maxAngularAcceleration": 180.0
-  },
-  "goalEndState": {
-    "velocity": 0,
-    "rotation": 0.0,
-    "rotateFast": true
-  },
-  "reversed": false,
-  "folder": "Amp Blue Side",
-  "previewStartingState": {
-    "rotation": 40.914383220025044,
-    "velocity": 0
-  },
-  "useDefaultConstraints": true
-}
