--- src/main/deploy/pathplanner/paths/Amp Side Blue trajectory2.path
@@ -1,87 +1,0 @@
-{
-  "version": 1.0,
-  "waypoints": [
-    {
-      "anchor": {
-        "x": 2.6933948404713925,
-        "y": 6.980826883268969
-      },
-      "prevControl": null,
-      "nextControl": {
-        "x": 1.8267774427286747,
-        "y": 7.180210827658122
-      },
-      "isLocked": false,
-      "linkedName": null
-    },
-    {
-      "anchor": {
-        "x": 1.136822330624718,
-        "y": 6.385008325572207
-      },
-      "prevControl": {
-        "x": 1.5950609668169533,
-        "y": 6.7145501533019445
-      },
-      "nextControl": null,
-      "isLocked": false,
-      "linkedName": null
-    }
-  ],
-  "rotationTargets": [],
-  "constraintZones": [],
-  "eventMarkers": [
-    {
-      "name": "New Event Marker",
-      "waypointRelativePos": 0.65,
-      "command": {
-        "type": "sequential",
-        "data": {
-          "commands": [
-            {
-              "type": "named",
-              "data": {
-                "name": "TransitShootSpeaker"
-              }
-            }
-          ]
-        }
-      }
-    },
-    {
-      "name": "New Event Marker",
-      "waypointRelativePos": 0.15,
-      "command": {
-        "type": "sequential",
-        "data": {
-          "commands": [
-            {
-              "type": "named",
-              "data": {
-                "name": "IntakeTransitStopCommand"
-              }
-            }
-          ]
-        }
-      }
-    }
-  ],
-  "globalConstraints": {
-    "maxVelocity": 2.0,
-    "maxAcceleration": 3.0,
-    "maxAngularVelocity": 360.0,
-    "maxAngularAcceleration": 180.0
-  },
-  "goalEndState": {
-    "velocity": 0,
-    "rotation": 41.347777219693604,
-    "rotateFast": true
-  },
-  "reversed": false,
-  "folder": "Amp Blue Side",
-  "previewStartingState": {
-    "rotation": -0.80176384086368,
-    "velocity": 0.0
-  },
-  "useDefaultConstraints": false
-}
