--- src/main/deploy/pathplanner/paths/Amp Side trajectory4.path
@@ -1,87 +1,0 @@
-{
-  "version": 1.0,
-  "waypoints": [
-    {
-      "anchor": {
-        "x": 7.6621605095062035,
-        "y": 7.320540680967402
-      },
-      "prevControl": null,
-      "nextControl": {
-        "x": 3.545818145767346,
-        "y": 6.127736927838529
-      },
-      "isLocked": false,
-      "linkedName": null
-    },
-    {
-      "anchor": {
-        "x": 1.171904793952038,
-        "y": 6.3616200166873265
-      },
-      "prevControl": {
-        "x": 1.5110352727827958,
-        "y": 6.876162812154685
-      },
-      "nextControl": null,
-      "isLocked": false,
-      "linkedName": null
-    }
-  ],
-  "rotationTargets": [],
-  "constraintZones": [],
-  "eventMarkers": [
-    {
-      "name": "New Event Marker",
-      "waypointRelativePos": 0.7,
-      "command": {
-        "type": "sequential",
-        "data": {
-          "commands": [
-            {
-              "type": "named",
-              "data": {
-                "name": "TransitShootSpeaker"
-              }
-            }
-          ]
-        }
-      }
-    },
-    {
-      "name": "New Event Marker",
-      "waypointRelativePos": 0.3,
-      "command": {
-        "type": "sequential",
-        "data": {
-          "commands": [
-            {
-              "type": "named",
-              "data": {
-                "name": "IntakeTransitStopCommand"
-              }
-            }
-          ]
-        }
-      }
-    }
-  ],
-  "globalConstraints": {
-    "maxVelocity": 6.0,
-    "maxAcceleration": 4.5,
-    "maxAngularVelocity": 360.0,
-    "maxAngularAcceleration": 180.0
-  },
-  "goalEndState": {
-    "velocity": 0,
-    "rotation": 33.231711067979354,
-    "rotateFast": true
-  },
-  "reversed": false,
-  "folder": "Amp Blue Side",
-  "previewStartingState": {
-    "rotation": 1.6365770416167964,
-    "velocity": 0
-  },
-  "useDefaultConstraints": false
-}
