--- src/main/deploy/pathplanner/paths/Blue Auto Trap.path
@@ -1,52 +1,0 @@
-{
-  "version": 1.0,
-  "waypoints": [
-    {
-      "anchor": {
-        "x": 14.862776074000834,
-        "y": 5.298802196649739
-      },
-      "prevControl": null,
-      "nextControl": {
-        "x": 14.762776074000834,
-        "y": 5.298802196649739
-      },
-      "isLocked": false,
-      "linkedName": null
-    },
-    {
-      "anchor": {
-        "x": 15.158847310534718,
-        "y": 5.417230691263294
-      },
-      "prevControl": {
-        "x": 15.258847310534719,
-        "y": 5.417230691263294
-      },
-      "nextControl": null,
-      "isLocked": false,
-      "linkedName": null
-    }
-  ],
-  "rotationTargets": [],
-  "constraintZones": [],
-  "eventMarkers": [],
-  "globalConstraints": {
-    "maxVelocity": 1.0,
-    "maxAcceleration": 3.0,
-    "maxAngularVelocity": 360.0,
-    "maxAngularAcceleration": 180.0
-  },
-  "goalEndState": {
-    "velocity": 0,
-    "rotation": 180.0,
-    "rotateFast": false
-  },
-  "reversed": false,
-  "folder": null,
-  "previewStartingState": {
-    "rotation": 180.0,
-    "velocity": 0
-  },
-  "useDefaultConstraints": false
-}
