--- src/main/deploy/pathplanner/paths/Custom 2note Non amp 1.path
@@ -1,52 +1,0 @@
-{
-  "version": 1.0,
-  "waypoints": [
-    {
-      "anchor": {
-        "x": 0.8289994622948672,
-        "y": 4.232945745127768
-      },
-      "prevControl": null,
-      "nextControl": {
-        "x": 1.1824781936076678,
-        "y": 4.2175771046359065
-      },
-      "isLocked": false,
-      "linkedName": null
-    },
-    {
-      "anchor": {
-        "x": 2.7633315409828914,
-        "y": 4.114517250514216
-      },
-      "prevControl": {
-        "x": 1.7961655016388791,
-        "y": 4.035564920771847
-      },
-      "nextControl": null,
-      "isLocked": false,
-      "linkedName": null
-    }
-  ],
-  "rotationTargets": [],
-  "constraintZones": [],
-  "eventMarkers": [],
-  "globalConstraints": {
-    "maxVelocity": 2.0,
-    "maxAcceleration": 3.0,
-    "maxAngularVelocity": 360.0,
-    "maxAngularAcceleration": 180.0
-  },
-  "goalEndState": {
-    "velocity": 0,
-    "rotation": 0,
-    "rotateFast": true
-  },
-  "reversed": false,
-  "folder": null,
-  "previewStartingState": {
-    "rotation": -60.255118703057725,
-    "velocity": 0
-  },
-  "useDefaultConstraints": false
-}
