--- src/main/deploy/pathplanner/paths/Middle Trajectory3.path
@@ -1,86 +1,0 @@
-{
-  "version": 1.0,
-  "waypoints": [
-    {
-      "anchor": {
-        "x": 1.32,
-        "y": 5.5
-      },
-      "prevControl": null,
-      "nextControl": {
-        "x": 3.3704058291307457,
-        "y": 5.928936302317049
-      },
-      "isLocked": false,
-      "linkedName": null
-    },
-    {
-      "anchor": {
-        "x": 7.135923559596407,
-        "y": 5.89385383898973
-      },
-      "prevControl": {
-        "x": 4.960810833302578,
-        "y": 7.051575128791284
-      },
-      "nextControl": {
-        "x": 7.399108863726596,
-        "y": 5.753771338404307
-      },
-      "isLocked": false,
-      "linkedName": null
-    },
-    {
-      "anchor": {
-        "x": 8.445668857149679,
-        "y": 5.776912294565331
-      },
-      "prevControl": {
-        "x": 7.90722955135736,
-        "y": 5.776912294565331
-      },
-      "nextControl": null,
-      "isLocked": false,
-      "linkedName": null
-    }
-  ],
-  "rotationTargets": [],
-  "constraintZones": [],
-  "eventMarkers": [
-    {
-      "name": "New Event Marker",
-      "waypointRelativePos": 0.75,
-      "command": {
-        "type": "sequential",
-        "data": {
-          "commands": [
-            {
-              "type": "named",
-              "data": {
-                "name": "IntakeTransitCommand"
-              }
-            }
-          ]
-        }
-      }
-    }
-  ],
-  "globalConstraints": {
-    "maxVelocity": 5.0,
-    "maxAcceleration": 2.0,
-    "maxAngularVelocity": 360.0,
-    "maxAngularAcceleration": 180.0
-  },
-  "goalEndState": {
-    "velocity": 0,
-    "rotation": 0.0,
-    "rotateFast": false
-  },
-  "reversed": false,
-  "folder": "Middle Auto Blue OLD",
-  "previewStartingState": {
-    "rotation": -0.34423885703228213,
-    "velocity": 0
-  },
-  "useDefaultConstraints": false
-}
