--- src/main/deploy/pathplanner/paths/MiddleTrajectory2.path
@@ -1,87 +1,0 @@
-{
-  "version": 1.0,
-  "waypoints": [
-    {
-      "anchor": {
-        "x": 2.9026396514331485,
-        "y": 5.5430292057165325
-      },
-      "prevControl": null,
-      "nextControl": {
-        "x": 3.1526396514331485,
-        "y": 5.5430292057165325
-      },
-      "isLocked": false,
-      "linkedName": "middleAutoWaypoint1"
-    },
-    {
-      "anchor": {
-        "x": 1.2654580294915574,
-        "y": 5.5430292057165325
-      },
-      "prevControl": {
-        "x": 1.0154580294915574,
-        "y": 5.5430292057165325
-      },
-      "nextControl": null,
-      "isLocked": false,
-      "linkedName": "middleAutoStarting"
-    }
-  ],
-  "rotationTargets": [],
-  "constraintZones": [],
-  "eventMarkers": [
-    {
-      "name": "New Event Marker",
-      "waypointRelativePos": 0.55,
-      "command": {
-        "type": "sequential",
-        "data": {
-          "commands": [
-            {
-              "type": "named",
-              "data": {
-                "name": "TransitShootSpeaker"
-              }
-            }
-          ]
-        }
-      }
-    },
-    {
-      "name": "New Event Marker",
-      "waypointRelativePos": 0.4,
-      "command": {
-        "type": "sequential",
-        "data": {
-          "commands": [
-            {
-              "type": "named",
-              "data": {
-                "name": "IntakeTransitStopCommand"
-              }
-            }
-          ]
-        }
-      }
-    }
-  ],
-  "globalConstraints": {
-    "maxVelocity": 6.0,
-    "maxAcceleration": 3.0,
-    "maxAngularVelocity": 360.0,
-    "maxAngularAcceleration": 180.0
-  },
-  "goalEndState": {
-    "velocity": 0,
-    "rotation": -0.525356301900408,
-    "rotateFast": false
-  },
-  "reversed": false,
-  "folder": "Middle Auto Blue OLD",
-  "previewStartingState": {
-    "rotation": 1.1227365357992924,
-    "velocity": 0
-  },
-  "useDefaultConstraints": true
-}
