--- src/main/deploy/pathplanner/paths/MiddleTrajectory4.path
@@ -1,109 +1,0 @@
-{
-  "version": 1.0,
-  "waypoints": [
-    {
-      "anchor": {
-        "x": 7.568607273966684,
-        "y": 5.87046553010485
-      },
-      "prevControl": null,
-      "nextControl": {
-        "x": 6.45766260193489,
-        "y": 7.004798511021523
-      },
-      "isLocked": false,
-      "linkedName": null
-    },
-    {
-      "anchor": {
-        "x": 2.1103565174320065,
-        "y": 5.73013567679557
-      },
-      "prevControl": {
-        "x": 3.6042889179795448,
-        "y": 6.3616200166873265
-      },
-      "nextControl": {
-        "x": 1.8104204502531145,
-        "y": 5.603352879044631
-      },
-      "isLocked": false,
-      "linkedName": null
-    },
-    {
-      "anchor": {
-        "x": 1.3940937283583965,
-        "y": 5.5196408968316515
-      },
-      "prevControl": {
-        "x": 1.3002697333664692,
-        "y": 5.485042371375285
-      },
-      "nextControl": null,
-      "isLocked": false,
-      "linkedName": null
-    }
-  ],
-  "rotationTargets": [
-    {
-      "waypointRelativePos": 1.2,
-      "rotationDegrees": 0.0,
-      "rotateFast": true
-    }
-  ],
-  "constraintZones": [],
-  "eventMarkers": [
-    {
-      "name": "New Event Marker",
-      "waypointRelativePos": 0.95,
-      "command": {
-        "type": "sequential",
-        "data": {
-          "commands": [
-            {
-              "type": "named",
-              "data": {
-                "name": "TransitShootSpeaker"
-              }
-            }
-          ]
-        }
-      }
-    },
-    {
-      "name": "New Event Marker",
-      "waypointRelativePos": 0.5,
-      "command": {
-        "type": "sequential",
-        "data": {
-          "commands": [
-            {
-              "type": "named",
-              "data": {
-                "name": "IntakeTransitStopCommand"
-              }
-            }
-          ]
-        }
-      }
-    }
-  ],
-  "globalConstraints": {
-    "maxVelocity": 6.0,
-    "maxAcceleration": 4.0,
-    "maxAngularVelocity": 360.0,
-    "maxAngularAcceleration": 180.0
-  },
-  "goalEndState": {
-    "velocity": 0,
-    "rotation": 0.857140600362757,
-    "rotateFast": true
-  },
-  "reversed": false,
-  "folder": "Middle Auto Blue OLD",
-  "previewStartingState": {
-    "rotation": 0.0,
-    "velocity": 0
-  },
-  "useDefaultConstraints": false
-}
