--- src/main/deploy/pathplanner/paths/New Non-Amp Blue Side trajectory 0.5.path
@@ -1,87 +1,0 @@
-{
-  "version": 1.0,
-  "waypoints": [
-    {
-      "anchor": {
-        "x": 2.47,
-        "y": 4.1046482092964185
-      },
-      "prevControl": null,
-      "nextControl": {
-        "x": 1.101783930234527,
-        "y": 4.3385312981452175
-      },
-      "isLocked": false,
-      "linkedName": null
-    },
-    {
-      "anchor": {
-        "x": 1.0066866689552871,
-        "y": 4.5531370872718
-      },
-      "prevControl": {
-        "x": 1.802804405061245,
-        "y": 4.425056694300723
-      },
-      "nextControl": null,
-      "isLocked": false,
-      "linkedName": null
-    }
-  ],
-  "rotationTargets": [],
-  "constraintZones": [],
-  "eventMarkers": [
-    {
-      "name": "New Event Marker",
-      "waypointRelativePos": 0.8,
-      "command": {
-        "type": "sequential",
-        "data": {
-          "commands": [
-            {
-              "type": "named",
-              "data": {
-                "name": "TransitShootSpeaker"
-              }
-            }
-          ]
-        }
-      }
-    },
-    {
-      "name": "New Event Marker",
-      "waypointRelativePos": 0.25,
-      "command": {
-        "type": "sequential",
-        "data": {
-          "commands": [
-            {
-              "type": "named",
-              "data": {
-                "name": "IntakeTransitStopCommand"
-              }
-            }
-          ]
-        }
-      }
-    }
-  ],
-  "globalConstraints": {
-    "maxVelocity": 2.0,
-    "maxAcceleration": 1.75,
-    "maxAngularVelocity": 360.0,
-    "maxAngularAcceleration": 180.0
-  },
-  "goalEndState": {
-    "velocity": 0,
-    "rotation": -39.42451665057289,
-    "rotateFast": true
-  },
-  "reversed": false,
-  "folder": "Non-Amp Side Blue",
-  "previewStartingState": {
-    "rotation": 0,
-    "velocity": 0
-  },
-  "useDefaultConstraints": false
-}
