--- src/main/deploy/pathplanner/paths/Non-Amp Blue Side trajectory 0.3.path
@@ -1,76 +1,0 @@
-{
-  "version": 1.0,
-  "waypoints": [
-    {
-      "anchor": {
-        "x": 1.183598948394478,
-        "y": 4.432084533684738
-      },
-      "prevControl": null,
-      "nextControl": {
-        "x": 2.1308254582321133,
-        "y": 4.046177437084219
-      },
-      "isLocked": false,
-      "linkedName": null
-    },
-    {
-      "anchor": {
-        "x": 2.8,
-        "y": 4.1046482092964185
-      },
-      "prevControl": {
-        "x": 1.7999999999999998,
-        "y": 4.1046482092964185
-      },
-      "nextControl": null,
-      "isLocked": false,
-      "linkedName": null
-    }
-  ],
-  "rotationTargets": [
-    {
-      "waypointRelativePos": 0.5,
-      "rotationDegrees": 0,
-      "rotateFast": false
-    }
-  ],
-  "constraintZones": [],
-  "eventMarkers": [
-    {
-      "name": "New Event Marker",
-      "waypointRelativePos": 0.35,
-      "command": {
-        "type": "sequential",
-        "data": {
-          "commands": [
-            {
-              "type": "named",
-              "data": {
-                "name": "IntakeTransitCommand"
-              }
-            }
-          ]
-        }
-      }
-    }
-  ],
-  "globalConstraints": {
-    "maxVelocity": 2.0,
-    "maxAcceleration": 1.5,
-    "maxAngularVelocity": 360.0,
-    "maxAngularAcceleration": 180.0
-  },
-  "goalEndState": {
-    "velocity": 0,
-    "rotation": 0,
-    "rotateFast": true
-  },
-  "reversed": false,
-  "folder": "Non-Amp Side Blue",
-  "previewStartingState": {
-    "rotation": -60.0,
-    "velocity": 0
-  },
-  "useDefaultConstraints": false
-}
