--- src/main/deploy/pathplanner/paths/Non-Amp Blue Side trajectory 1.path
@@ -1,81 +1,0 @@
-{
-  "version": 1.0,
-  "waypoints": [
-    {
-      "anchor": {
-        "x": 0.9146333962183594,
-        "y": 4.572414386994017
-      },
-      "prevControl": null,
-      "nextControl": {
-        "x": 3.404470684154232,
-        "y": 1.829576319779909
-      },
-      "isLocked": false,
-      "linkedName": null
-    },
-    {
-      "anchor": {
-        "x": 8.309255573380506,
-        "y": 0.7897304155653457
-      },
-      "prevControl": {
-        "x": 2.037164437274244,
-        "y": 0.6832144578778497
-      },
-      "nextControl": null,
-      "isLocked": false,
-      "linkedName": null
-    }
-  ],
-  "rotationTargets": [
-    {
-      "waypointRelativePos": 0.2,
-      "rotationDegrees": -127.12902284000809,
-      "rotateFast": true
-    },
-    {
-      "waypointRelativePos": 0.7,
-      "rotationDegrees": 0.0,
-      "rotateFast": true
-    }
-  ],
-  "constraintZones": [],
-  "eventMarkers": [
-    {
-      "name": "New Event Marker",
-      "waypointRelativePos": 0.45,
-      "command": {
-        "type": "sequential",
-        "data": {
-          "commands": [
-            {
-              "type": "named",
-              "data": {
-                "name": "IntakeTransitCommand"
-              }
-            }
-          ]
-        }
-      }
-    }
-  ],
-  "globalConstraints": {
-    "maxVelocity": 4.0,
-    "maxAcceleration": 3.5,
-    "maxAngularVelocity": 360.0,
-    "maxAngularAcceleration": 180.0
-  },
-  "goalEndState": {
-    "velocity": 0,
-    "rotation": 0.0,
-    "rotateFast": false
-  },
-  "reversed": false,
-  "folder": "Non-Amp Side Blue",
-  "previewStartingState": {
-    "rotation": -57.739857024931524,
-    "velocity": 0.0
-  },
-  "useDefaultConstraints": false
-}
