+++ src/main/java/frc/robot/subsystems/drive/swerve/SwerveUtils.java
@@ -1,4 +1,4 @@
-package frc.robot.swerve;
+package frc.robot.subsystems.drive.swerve;
 
 public class SwerveUtils {
